+++ index.php
@@ -15,178 +15,7 @@
     $isAuth = false;
   }
 
-  $navLinks = [
-    array(
-      'Name' => 'Home',
-      'Title' => 'Home',
-      'ACL' => null,
-      'Type' => 'Link', // Link / MenuLink / SubMenuLink / Menu / Submenu
-      'Menu' => null,
-      'Submenu' => null,
-      'Url' => '#page=core/default',
-      'Icon' => 'fa fa-house'
-    ),
-    array(
-      'Name' => 'DNS Toolbox',
-      'Title' => 'DNS Toolbox',
-      'ACL' => 'DNS-TOOLBOX',
-      'Type' => 'Link', // Link / MenuLink / SubMenuLink / Menu / Submenu
-      'Menu' => null,
-      'Submenu' => null,
-      'Url' => '#page=tools/dnstoolbox',
-      'Icon' => 'fa fa-toolbox'
-    ),
-    array(
-      'Name' => 'Security Assessment',
-      'Title' => 'Security Assessment Report Generator',
-      'ACL' => 'B1-SECURITY-ASSESSMENT',
-      'Type' => 'Link',
-      'Menu' => null,
-      'Submenu' => null,
-      'Url' => '#page=plugin/IB/security-assessment-generator',
-      'Icon' => 'fa fa-magnifying-glass-chart'
-    ),
-    array(
-      'Name' => 'Threat Actors',
-      'Title' => 'Threat Actors',
-      'ACL' => 'B1-THREAT-ACTORS',
-      'Type' => 'Link',
-      'Menu' => null,
-      'Submenu' => null,
-      'Url' => '#page=plugin/IB/threat-actors',
-      'Icon' => 'fa fa-skull'
-    ),
-    array(
-      'Name' => 'Dev',
-      'Title' => 'Dev',
-      'ACL' => null,
-      'Type' => 'Menu',
-      'Menu' => null,
-      'Submenu' => null,
-      'Url' => null,
-      'Icon' => 'fa fa-toolbox'
-    ),
-    array(
-      'Name' => 'License Utilization',
-      'Title' => 'License Utilization',
-      'ACL' => 'B1-LICENSE-USAGE',
-      'Type' => 'MenuLink',
-      'Menu' => 'Dev',
-      'Submenu' => null,
-      'Url' => '#page=plugin/IB/license-usage',
-      'Icon' => 'fas fa-certificate'
-    ),
-    array(
-      'Name' => 'Admin',
-      'Title' => 'Admin',
-      'ACL' => 'ADMIN-Menu',
-      'Type' => 'Menu',
-      'Menu' => null,
-      'Submenu' => null,
-      'Url' => null,
-      'Icon' => 'fas fa-user-shield'
-    ),
-    array(
-      'Name' => 'Settings',
-      'Title' => 'Settings',
-      'ACL' => null,
-      'Type' => 'SubMenu',
-      'Menu' => 'Admin',
-      'Submenu' => null,
-      'Url' => null,
-      'Icon' => 'fa fa-cog'
-    ),
-    array(
-      'Name' => 'Users',
-      'Title' => 'Users',
-      'ACL' => 'ADMIN-USERS',
-      'Type' => 'SubMenuLink',
-      'Menu' => 'Admin',
-      'Submenu' => 'Settings',
-      'Url' => '#page=core/users',
-      'Icon' => null
-    ),
-    array(
-      'Name' => 'Configuration',
-      'Title' => 'Configuration',
-      'ACL' => 'ADMIN-CONFIG',
-      'Type' => 'SubMenuLink',
-      'Menu' => 'Admin',
-      'Submenu' => 'Settings',
-      'Url' => '#page=core/configuration',
-      'Icon' => null
-    ),
-    array(
-      'Name' => 'Role Based Access',
-      'Title' => 'Role Based Access',
-      'ACL' => 'ADMIN-RBAC',
-      'Type' => 'SubMenuLink',
-      'Menu' => 'Admin',
-      'Submenu' => 'Settings',
-      'Url' => '#page=core/rbac',
-      'Icon' => null
-    ),
-    array(
-      'Name' => 'Security Assessment',
-      'Title' => 'Security Assessment',
-      'ACL' => 'ADMIN-SECASS',
-      'Type' => 'SubMenuLink',
-      'Menu' => 'Admin',
-      'Submenu' => 'Settings',
-      'Url' => '#page=plugin/IB/security-assessment-configuration',
-      'Icon' => null
-    ),
-    array(
-      'Name' => 'Logs',
-      'Title' => 'Logs',
-      'ACL' => null,
-      'Type' => 'SubMenu',
-      'Menu' => 'Admin',
-      'Submenu' => null,
-      'Url' => null,
-      'Icon' => 'fa-regular fa-file'
-    ),
-    array(
-      'Name' => 'Portal Logs',
-      'Title' => 'Logs',
-      'ACL' => 'ADMIN-LOGS',
-      'Type' => 'SubMenuLink',
-      'Menu' => 'Admin',
-      'Submenu' => 'Logs',
-      'Url' => '#page=core/logs',
-      'Icon' => null
-    ),
-    array(
-      'Name' => 'Reports',
-      'Title' => 'Reports',
-      'ACL' => 'REPORT-Menu',
-      'Type' => 'Menu',
-      'Menu' => null,
-      'Submenu' => null,
-      'Url' => null,
-      'Icon' => 'fa-solid fa-chart-simple'
-    ),
-    array(
-      'Name' => 'Assessments',
-      'Title' => 'Assessment Reporting',
-      'ACL' => 'REPORT-ASSESSMENTS',
-      'Type' => 'MenuLink',
-      'Menu' => 'Reports',
-      'Submenu' => null,
-      'Url' => '#page=plugin/IB/assessment-reporting',
-      'Icon' => 'fa-solid fa-arrows-to-eye'
-    ),
-    array(
-      'Name' => 'Web Tracking',
-      'Title' => 'Web Tracking',
-      'ACL' => 'REPORT-TRACKING',
-      'Type' => 'MenuLink',
-      'Menu' => 'Reports',
-      'Submenu' => null,
-      'Url' => '#page=reports/tracking',
-      'Icon' => 'fa-solid fa-bullseye'
-    )
-  ];
+  $navLinks = $ib->pages->get();
 
   function filterNavLinksByMenu($navLinks, $menuName) {
     return array_filter($navLinks, function($link) use ($menuName) {
@@ -217,11 +46,11 @@
   <div class="sidebar">
     <div class="logo-details">
       <?php
-      $smLogoPath = $ib->config->getConfig('Styling', 'logo-sm');
+      $smLogoPath = $ib->config->get('Styling', 'logo-sm')['Image'];
       $smLogoPath = $smLogoPath ? $smLogoPath : '/assets/images/php-ef-icon.png';
       echo '<img class="logo-sm" src="' . (file_exists(__DIR__ . $smLogoPath) ? $smLogoPath : '/assets/images/php-ef-icon.png') . '"></img>';
 
-      $lgLogoPath = $ib->config->getConfig('Styling', 'logo-lg');
+      $lgLogoPath = $ib->config->get('Styling', 'logo-lg')['Image'];
       $lgLogoPath = $lgLogoPath ? $lgLogoPath : '/assets/images/php-ef-icon-text.png';
       echo '<img class="logo-lg" src="' . (file_exists(__DIR__ . $lgLogoPath) ? $lgLogoPath : '/assets/images/php-ef-icon-text.png') . '"></img>';
       ?>